Introduce MaxRateRegistry interface
